Detecting of the the internet connection fail
Currently I simulates a disconnecting of my internet cable after
successfull connection to a IRC server.
I want to see the socket error event and disconnect of the socket if:
1. Established connection was aborted by the software in your host
machine
2. Connection timeout. Connection attempt failed because the connected
party did not properly respond after a period of time, or established
connection failed because connected host has failed to respond.