user's channel modes
I think the IrcUser.ChannelModes value is incorrect. For example if you
send to the server the follow sequence of  commands:
MODE #bonalink +o anhr
anhr user is #bonalink channel operator
MODE #bonalink +v anhr
anhr user is channel operator and have voice in the #bonalink channel
But you can see only 'v' value in the IrcUser.ChannelModes[#bonalink].
The 'o' value is lost.
I have replace the char? to a List<char?> in the IrcUser.ChannelsModes
dictionary for resolving of issue.
Also if user is channel operator and have voice in the channel, then you
must see this user as channel operartor